# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a restaurant management system.",
        "title": "Restaurant Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/orderitems": {
            "get": {
                "description": "Get all order items from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderItems"
                ],
                "summary": "Get all order items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CustomBsonM"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/orderitems/order/{order_id}": {
            "get": {
                "description": "Get order items from the database by order ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderItems"
                ],
                "summary": "Get order items by order ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/orderitems/{order_item_id}": {
            "get": {
                "description": "Get a single order item from the database by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderItems"
                ],
                "summary": "Get a single order item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order Item ID",
                        "name": "order_item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrderItem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an order item in the database by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderItems"
                ],
                "summary": "Update an order item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order Item ID",
                        "name": "order_item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Order Item",
                        "name": "orderItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OrderItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UpdateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CustomBsonM": {
            "type": "object",
            "properties": {
                "field1": {
                    "type": "string"
                },
                "field2": {
                    "type": "string"
                }
            }
        },
...
        "model.OrderItem": {
            "description": "Order item details swagger:model",
            "type": "object",
            "required": [
                "food_id",
                "order_id",
                "quantity",
                "unit_price"
            ],
            "properties": {
                "created_at": {
                    "description": "Created timestamp of the order item",
                    "type": "string"
                },
                "food_id": {
                    "description": "ID of the associated food item\n@swag.Required",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the order item\n@swag.Type string",
                    "type": "string"
                },
                "order_id": {
                    "description": "ID of the associated order\n@swag.Required",
                    "type": "string"
                },
                "order_item_id": {
                    "description": "Unique ID of the order item",
                    "type": "string"
                },
                "quantity": {
                    "description": "Quantity of the order item\n@swag.Required",
                    "type": "string"
                },
                "unit_price": {
                    "description": "Unit price of the order item\n@swag.Required",
                    "type": "number"
                },
                "updated_at": {
                    "description": "Updated timestamp of the order item",
                    "type": "string"
                }
            }
        },
...
        "model.UpdateResult": {
            "type": "object",
            "properties": {
                "matchedCount": {
                    "type": "integer"
                },
                "modifiedCount": {
                    "type": "integer"
                },
                "upsertedCount": {
                    "type": "integer"
                },
                "upsertedId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}